Use TelemacToolsDockwidgetBase and promote widget as qgsdockwidget
