--- telemac_tools/telemac_tools_dockwidget_base.ui
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>TelemacToolsDockWidgetBase</class>
- <widget class="QDockWidget" name="TelemacToolsDockWidgetBase">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>232</width>
-    <height>141</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Telemac Tools</string>
-  </property>
-  <widget class="QWidget" name="dockWidgetContents">
-   <layout class="QGridLayout" name="gridLayout">
-    <item row="0" column="0">
-     <widget class="QLabel" name="label">
-      <property name="text">
-       <string>Replace this QLabel
-with the desired
-plugin content.</string>
-      </property>
-     </widget>
-    </item>
-   </layout>
-  </widget>
- </widget>
- <resources/>
- <connections/>
-</ui>
